Viele Tabs repariert
git-svn-id: https://callmonitor.svn.sourceforge.net/svnroot/callmonitor/fritzbox/callmonitor/trunk@61 e872857c-5b07-0410-aa81-e3586f851bf3

+++ base/usr/lib/callmonitor/lock.sh
@@ -1,14 +1,20 @@
 # lock $file by creating a symlink $file.lock -> PID
 lock() {
-	local file="$(realpath "$1")" interval="${2:-1000000}" first=1
+	local file="$1" interval="${2:-1000000}" first=1
+	# race conditions between touch and realpath still possible
+	if [ ! -e "$1" ] && ! touch "$1"; then
+		return 1
+	fi
+	file="$(realpath "$file")"
 	local lock="$file.lock"
 	while ! ln -s $$ "$lock" 2> /dev/null; do
-	if [ $first -eq 1 ]; then 
-		first=0
-		echo "Waiting for exclusive lock on $file" 1>&2
-	fi
-	usleep $interval
+		if [ $first -eq 1 ]; then 
+			first=0
+			echo "Waiting for exclusive lock on $file" 1>&2
+		fi
+		usleep $interval
 	done
+	return 0
 }
 
 unlock() {
